fix(coding): Do not `cmp` in command mode (#30)
This fixes a bug where when a user tried to `<tab>` complete something
in command mode, it wasn't working.

For example, if you wanted to run `:Telescope` but you typed in `:Tel`
and pressed `<TAB>`, it didn't attempt to auto-complete.

This is a really annoying bug because we aren't able to auto-complete
for plugins that have been loaded.