feat(legacy): disable mouse mode by default (#115)
In `vim_dotfiles` mouse mode is disabled by default (in tmux as well).
While pair programming and sharing a tmux for ops, etc, the navigator is
typically accustomed to being able to select text independently of the
cursor of the pair driving with the keyboard.

For that use-case: yes, it is true that by default one can disable the
mouse by holding shift.

I wonder if it would be possible to turn that on its' head and make it
easy to enable mouse mode situationally with shift?

Alternatively, and I am super open to just socializing knowledge about
the fact that folks can shift-select so long as that's not a barrier
to my team adopting.

This closes #95.