Dynamically size Telescope to balance context and pane size (#156)
In 44a5f70 / #48 we overrode the default Telescope layout from 80%
horizontal/90% vertical to be 100% both directions all the time. That
was a divergence from our existing vim_dotfiles setup (where fzf opens a
modal smaller in both directions than the window size) that happened
without a lot of discussion or our new habit of running polls for
divergence.

Several people have found the new behavior to be jarring and confusing.
In my personal experience I get broken out of context pretty frequently.
One good example: if you open either fuzzy-find or bundle-explorer from
the right hand pane in a two pane horizontal layout then the picker list
appears to take over the left hand pane (since we have a preview on the
right) which is confusing (why is it suddenly changing panes?). In
reality it's a modal, but there are no affordances at 100% sizing to
indicate that. In buffer-explorer the contrast with vim_dotfiles is more
stark, since there bundle-explorer opened up only in the pane you
triggered it from.

That being said: the original context for the change was that in small
enough windows having padding uses up valuable screen real estate.

Hence here I've tried to take a best-of-both-worlds approach: in decent
sized terminals we provide a base amount of padding (plus an amount that
scales with size) but switch into 100% mode when the window is small
enough.