Keep PageDown from scrolling further than necessary (#231)
This makes it stop in the middle of the screen when paging down further
would just result in only the final line of the file being down. The BT
vim experience just stops paging down when you hit the final line, so it
sometimes shows more or less of half a page, so this isn't 1 to 1... but
it's close enough, and more importantly, it's useful.

Slack convo:
[internal slack link removed]